src/aiops/requirements.txt
strands-agents[mcp]
fastapi
uvicorn[standard]
pydantic
nest_asyncio
langfuse
